refactor: 역할별 모듈로 이동
