+++ saerojinro-admin/src/main/java/goorm/saerojinro/admin/notification/application/NotificationAdminFacade.java
@@ -1,18 +1,16 @@
-package goorm.saerojinro.infra.notification.application;
+package goorm.saerojinro.admin.notification.application;
 
-import goorm.saerojinro.infra.notification.exception.EmitterNotFoundException;
-import goorm.saerojinro.infra.notification.request.NotificationSendRequest;
 import goorm.saerojinro.domain.notification.application.NotificationCommandService;
-import goorm.saerojinro.domain.notification.application.NotificationQueryService;
 import goorm.saerojinro.domain.notification.domain.EmitterRepository;
 import goorm.saerojinro.domain.notification.domain.Notification;
 import goorm.saerojinro.domain.reservation.application.ReservationQueryService;
 import goorm.saerojinro.domain.reservation.domain.Reservation;
 import goorm.saerojinro.domain.user.application.UserQueryService;
 import goorm.saerojinro.domain.user.domain.User;
-import goorm.saerojinro.infra.notification.response.ReceivedNotificationListResponse;
-import goorm.saerojinro.infra.notification.sse.NotificationSseSender;
+import goorm.saerojinro.infra.notification.exception.EmitterNotFoundException;
+import goorm.saerojinro.infra.notification.request.NotificationSendRequest;
 import goorm.saerojinro.infra.notification.response.NotificationSendResponse;
+import goorm.saerojinro.infra.notification.sse.NotificationSseSender;
 import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Component;
 import org.springframework.transaction.annotation.Transactional;
@@ -22,18 +20,12 @@ import java.util.List;
 
 @Component
 @RequiredArgsConstructor
-public class NotificationFacade {
+public class NotificationAdminFacade {
 	private final NotificationCommandService notificationCommandService;
-	private final NotificationQueryService notificationQueryService;
 	private final EmitterRepository emitterRepository;
 	private final UserQueryService userQueryService;
 	private final ReservationQueryService reservationQueryService;
 	private final NotificationSseSender sseSender;
-
-	public SseEmitter subscribe() {
-		User user = userQueryService.me();
-		return emitterRepository.save(user.getId());
-	}
 
 	@Transactional
 	public void sendNotificationByLectureId(Long lectureId, NotificationSendRequest request) {
@@ -75,12 +67,5 @@ public class NotificationFacade {
 		for (SseEmitter emitter : all) {
 			sseSender.sendNotification(emitter, message);
 		}
-	}
-
-	@Transactional(readOnly = true)
-	public ReceivedNotificationListResponse myNotification() {
-		User user = userQueryService.me();
-		List<Notification> notificationList = notificationQueryService.findByUserId(user.getId());
-		return ReceivedNotificationListResponse.from(notificationList);
 	}
 }
