+++ saerojinro-admin/src/main/java/goorm/saerojinro/admin/notification/presentation/NotificationAdminController.java
@@ -37,4 +37,17 @@ public interface NotificationAdminController {
 			required = true
 		) @Valid @RequestBody NotificationSendRequest sendRequest
 	);
+
+	@Operation(summary = "개별 알림 전송 API", description = """
+			- Description : 이 API는 특정 사용자에게 알림을 전송합니다
+			- Assignee : 이신행
+		""")
+	@ApiResponse(responseCode = "204")
+	ResponseEntity<Void> sendByUserId(
+		@PathVariable Long userId,
+		@Parameter(
+			description = "전체 알림 request 객체 입니다",
+			required = true
+		) @Valid @RequestBody NotificationSendRequest sendRequest
+	);
 }
